requirements.txt
discord.py
googletrans==4.0.0-rc1
fastapi
httpx
uvicorn